Move Lorewalker script from custom file, tabs
